nextclip and prevclip methods in.
I've got the nextclip and prevclip methods in. However the jumpreel
business is kicking up a bit of a stink due to some weird accidental
boolean coming in on the action.

I'll get that looked in to later when I'm not about to get off a train.
Wonder what it could be. S'weird anyway...

I'll figure it out.