Sprites only draw within Stage viewport
Scenes are kind-of sort-of in. They definitely need some kind of Lua
structure building up around them. Everything draws ok and looks fine
anyhow. If there's any weirdness I haven't come across it yet.

TODO
* I need to implement the Actor metamethods. __index and __newindex,
__index will be the table containing the nextclip, prevclip lot.
__newindex will try to match the key with the method names, and if it is
one it'll bail. This really needs doing before I can progress.
* Clicking on Actors (so Auras and the like) need to go in.
* Props will need in soon too, as I plan to use them as part of the debug
system. A lot of the logic can be recycled from Actors.
* Debug props: script loader, viewport changer, scene warper
* Gambit system will need to go in too, again I plan to use it for debug

I'm sure the existing Lime will be a source of a good chunk of the code.
Having to Lua-ify stuff might be a barrier. We'll see.

It's starting to look like a game again now. =D