Renamed 'lima'
This library was never officially 'lime'

Name changed to 'lima' to reflect this. Lima is portugese for lime,
as this framework is bundled up with Lua.

lime is a yet unreleased engine that is similar to this one.

I'm thinking of picking this one back up and porting it to SDL 2.